# app-service/api-docs/api-docs.yaml
openapi: 3.1.0
info:
  title: BiblioOnline API
  description: "API REST para la gestión de libros, préstamos, reseñas y usuarios."
  version: 1.0.0
servers:
- url: https://localhost:8443
  description: Generated server url
paths:
  /api/v1/users/{id}:
    get:
      tags:
      - user-rest-controller
      operationId: getUser
      parameters:
      - name: id
        in: path
        required: true
        schema:
          type: integer
          format: int64
      responses:
        "200":
          description: OK
          content:
            '*/*':
              schema:
                $ref: "#/components/schemas/UserDTO"
    put:
      tags:
      - user-rest-controller
      operationId: updateUser
      parameters:
      - name: id
        in: path
        required: true
        schema:
          type: integer
          format: int64
      requestBody:
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/UserUpdateDTO"
        required: true
      responses:
        "200":
          description: OK
          content:
            '*/*':
              schema:
                $ref: "#/components/schemas/UserDTO"
    delete:
      tags:
      - user-rest-controller
      operationId: deleteUser
      parameters:
      - name: id
        in: path
        required: true
        schema:
          type: integer
          format: int64
      responses:
        "200":
          description: OK
...
  /api/v1/books/{bookId}/reviews:
    post:
      tags:
      - review-rest-controller
      operationId: createReview
      parameters:
      - name: bookId
        in: path
        required: true
        schema:
          type: integer
          format: int64
      requestBody:
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/ReviewCreateDTO"
        required: true
      responses:
        "200":
          description: OK
          content:
            '*/*':
              schema:
                $ref: "#/components/schemas/ReviewDTO"
...
  /api/v1/reviews:
    get:
      tags:
      - review-rest-controller
      operationId: getReviews
      parameters:
      - name: q
        in: query
        required: false
        schema:
          type: string
      - name: page
        in: query
        required: false
        schema:
          type: integer
          format: int32
          default: 0
      - name: size
        in: query
        required: false
        schema:
          type: integer
          format: int32
          default: 7
      responses:
        "200":
          description: OK
          content:
            '*/*':
              schema:
                $ref: "#/components/schemas/PageReviewDTO"
  /api/v1/reviews/{id}:
    get:
      tags:
      - review-rest-controller
      operationId: getReview
      parameters:
      - name: id
        in: path
        required: true
        schema:
          type: integer
          format: int64
      responses:
        "200":
          description: OK
          content:
            '*/*':
              schema:
                $ref: "#/components/schemas/ReviewDTO"
    delete:
      tags:
      - review-rest-controller
      operationId: deleteReview
      parameters:
      - name: id
        in: path
        required: true
        schema:
          type: integer
          format: int64
      responses:
        "200":
          description: OK
...
  /api/v1/books/{id}/reviews:
    get:
      tags:
      - book-rest-controller
      operationId: getBookReviews
      parameters:
      - name: id
        in: path
        required: true
        schema:
          type: integer
          format: int64
      responses:
        "200":
          description: OK
          content:
            '*/*':
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/ReviewDTO"
components:
  schemas:
    UserUpdateDTO:
      type: object
      properties:
        name:
          type: string
        surname:
          type: string
        email:
          type: string
        description:
          type: string
    UserDTO:
      type: object
      properties:
        id:
          type: integer
          format: int64
        name:
          type: string
        surname:
          type: string
        email:
          type: string
        description:
          type: string
        imageUrl:
          type: string
...
    ReviewCreateDTO:
      type: object
      properties:
        comment:
          type: string
        rating:
          type: integer
          format: int32
    ReviewDTO:
      type: object
      properties:
        id:
          type: integer
          format: int64
        comment:
          type: string
        rating:
          type: integer
          format: int32
        userId:
          type: integer
          format: int64
        userName:
          type: string
        bookId:
          type: integer
          format: int64
        bookTitle:
          type: string
...
    PageableObject:
      type: object
      properties:
        offset:
          type: integer
          format: int64
        unpaged:
          type: boolean
        pageSize:
          type: integer
          format: int32
        pageNumber:
          type: integer
          format: int32
        sort:
          $ref: "#/components/schemas/SortObject"
        paged:
          type: boolean
    SortObject:
      type: object
      properties:
        empty:
          type: boolean
        unsorted:
          type: boolean
        sorted:
          type: boolean
    PageReviewDTO:
      type: object
      properties:
        totalPages:
          type: integer
          format: int32
        totalElements:
          type: integer
          format: int64
        size:
          type: integer
          format: int32
        content:
          type: array
          items:
            $ref: "#/components/schemas/ReviewDTO"
        number:
          type: integer
          format: int32
        first:
          type: boolean
        last:
          type: boolean
        numberOfElements:
          type: integer
          format: int32
        pageable:
          $ref: "#/components/schemas/PageableObject"
        sort:
          $ref: "#/components/schemas/SortObject"
        empty:
          type: boolean
...